# swagger.yaml
openapi: 3.0.3
info:
  title: Branded UK E-Commerce API
  description: |
    # Overview
    This API provides endpoints for the Branded UK e-commerce platform.
    
    ## Base URL
    **Production**: `https://api.brandeduk.com`
    
    ## Authentication
    No authentication required. Public API.
    
    ## Endpoint Categories
    
    ### Products (Main)
    - `GET /api/products` - Main product listing with combined filters
    - `GET /api/products/filters` - Product type counts for filtering
    - `GET /api/products/types` - All product types with counts
    - `GET /api/products/{code}` - Single product details
    - `GET /api/products/{code}/pricing` - Get pricing information for a style code
    
    ### Contact & Quotes
    - `POST /api/contact` - Submit contact form inquiry
    - `POST /api/quotes` - Submit quote request with optional logo file uploads
    
    ### Filter Endpoints (Testing & Database Exploration)
    Each filter type has TWO endpoints:
    1. **List all values**: Returns all options from that database table
    2. **Get products by slug**: Returns products matching a specific filter value
    
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.brandeduk.com
    description: Production Server
  - url: http://localhost:3004
    description: Local Development
...
paths:
...
  /api/pricing/rules:
    get:
      tags:
        - Pricing
      summary: List pricing rules
      description: |
        Returns all pricing rules. You can optionally filter to only active rules.
      operationId: listPricingRules
      parameters:
        - name: active
          in: query
          required: false
          description: If "true", only return active rules
          schema:
            type: string
            enum: ["true","false"]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        version:
                          type: string
                          example: "1.0"
                        from_price:
                          type: number
                          example: 10.00
                        to_price:
                          type: number
                          example: 14.99
                        markup_percent:
                          type: number
                          example: 127.00
                        active:
                          type: boolean
                          example: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Pricing
      summary: Create a new pricing rule
      operationId: createPricingRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from_price
                - to_price
                - markup_percent
              properties:
                version:
                  type: string
                  description: Version label for this rule set
                  default: "1.0"
                  example: "1.0"
                from_price:
                  type: number
                  example: 10.00
                to_price:
                  type: number
                  example: 14.99
                markup_percent:
                  type: number
                  example: 127.00
                active:
                  type: boolean
                  description: Whether this rule is active
                  default: false
      responses:
        '201':
          description: Pricing rule created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      version:
                        type: string
                      from_price:
                        type: number
                      to_price:
                        type: number
                      markup_percent:
                        type: number
                      active:
                        type: boolean
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pricing/rules/{id}:
    put:
      tags:
        - Pricing
      summary: Update an existing pricing rule
      description: |
        Update a pricing rule identified by a composite key.

        You can either:
        - Pass `id` in the path as `version:from_price:to_price`, e.g. `1.0:10.00:14.99`, or
        - Provide the original key in the body via `originalVersion`, `originalFromPrice`, `originalToPrice`.
      operationId: updatePricingRule
      parameters:
        - name: id
          in: path
          required: true
          description: Composite key in the format `version:from_price:to_price`
          schema:
            type: string
            example: "1.0:10.00:14.99"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: string
                  description: New version value (optional)
                from_price:
                  type: number
                  description: New from_price (optional)
                to_price:
                  type: number
                  description: New to_price (optional)
                markup_percent:
                  type: number
                  description: New markup percentage (optional)
                active:
                  type: boolean
                  description: New active flag (optional)
                originalVersion:
                  type: string
                  description: Original version key (optional)
                originalFromPrice:
                  type: number
                  description: Original from_price key (optional)
                originalToPrice:
                  type: number
                  description: Original to_price key (optional)
      responses:
        '200':
          description: Pricing rule updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      version:
                        type: string
                      from_price:
                        type: number
                      to_price:
                        type: number
                      markup_percent:
                        type: number
                      active:
                        type: boolean
        '400':
          description: Validation error (missing key or fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Pricing rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pricing/reprice:
    post:
      tags:
        - Pricing
      summary: Recalculate sell_price for all products
      description: |
        Recalculates `sell_price` for all products using the currently active pricing rules
        and refreshes the `product_search_materialized` view.
      operationId: repriceProducts
      responses:
        '200':
          description: Repricing job completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Repricing job completed"
                  updatedProducts:
                    type: integer
                    description: Number of products updated
                  materializedViewRefreshed:
                    type: boolean
                  refreshError:
                    type: string
                    nullable: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pricing/products/{code}/sell-price:
    put:
      tags:
        - Pricing
      summary: Direct sell_price updates are disabled
      description: |
        This endpoint is deprecated. Direct `sell_price` updates are no longer allowed.
        Use the admin carton price endpoint instead:
        `PUT /api/admin/products/{code}/carton-price`, which recalculates `sell_price`
        from the current pricing rules.
      operationId: overrideSellPrice
      parameters:
        - name: code
          in: path
          required: true
          description: Product style code (e.g. JC001)
          schema:
            type: string
            example: "JC001"
      requestBody:
        required: false
      responses:
        '400':
          description: Direct sell_price updates are disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # ADMIN ENDPOINTS
  # ============================================================================
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...